# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cloud Infrastructure Optimizer",
    "description": "Local capstone demo. Mock data and simulated actions; authentication is not implemented.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/resources": {
      "get": {
        "tags": [
          "Resources"
        ],
        "summary": "Resources",
        "operationId": "resources_api_v1_resources_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ec2",
                    "rds"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ResourceResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recommendations": {
      "get": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Recommendations",
        "operationId": "recommendations_api_v1_recommendations_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "open",
                    "approved",
                    "dismissed",
                    "resolved"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Recommendation_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/executions": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "Executions",
        "operationId": "executions_api_v1_executions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExecutionJob_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecutionJob": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recommendation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recommendation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "recommendation_id",
          "status",
          "message",
          "created_at"
        ],
        "title": "ExecutionJob"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Metric": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "aggregation": {
            "type": "string",
            "enum": [
              "average",
              "maximum",
              "minimum",
              "sum"
            ],
            "title": "Aggregation"
          },
          "sample_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sample Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "unit",
          "aggregation",
          "sample_count"
        ],
        "title": "Metric"
      },
...
      "Page_ExecutionJob_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExecutionJob"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "data_source": {
            "type": "string",
            "const": "mock",
            "title": "Data Source",
            "default": "mock"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ExecutionJob]"
      },
      "Page_Recommendation_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "data_source": {
            "type": "string",
            "const": "mock",
            "title": "Data Source",
            "default": "mock"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Recommendation]"
      },
      "Page_ResourceResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResourceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "data_source": {
            "type": "string",
            "const": "mock",
            "title": "Data Source",
            "default": "mock"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ResourceResponse]"
      },
      "Recommendation": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "finding_type": {
            "type": "string",
            "enum": [
              "idle_ec2",
              "oversized_rds",
              "cost_anomaly"
            ],
            "title": "Finding Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          },
          "proposed_action": {
            "type": "string",
            "enum": [
              "stop_ec2",
              "review_rds",
              "investigate_cost"
            ],
            "title": "Proposed Action"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Level"
          },
          "estimated_monthly_savings": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Monthly Savings"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "USD"
          },
          "savings_assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Savings Assumptions"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "approved",
              "dismissed",
              "resolved"
            ],
            "title": "Status",
            "default": "open"
          },
          "data_source": {
            "type": "string",
            "enum": [
              "mock",
              "aws"
            ],
            "title": "Data Source",
            "default": "mock"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "explanation_source": {
            "type": "string",
            "enum": [
              "mock",
              "claude",
              "rules_fallback"
            ],
            "title": "Explanation Source",
            "default": "mock"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "finding_type",
          "title",
          "explanation",
          "evidence",
          "proposed_action",
          "risk_level",
          "created_at"
        ],
        "title": "Recommendation"
      },
      "ResourceResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "resource_type": {
            "type": "string",
            "enum": [
              "ec2",
              "rds"
            ],
            "title": "Resource Type"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Window End"
          },
          "configuration": {
            "additionalProperties": true,
            "type": "object",
            "title": "Configuration"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags"
          },
          "metrics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Metric"
            },
            "type": "object",
            "title": "Metrics"
          },
          "data_quality": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "insufficient"
            ],
            "title": "Data Quality"
          },
          "missing_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Metrics"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "data_source": {
            "type": "string",
            "const": "mock",
            "title": "Data Source",
            "default": "mock"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resource_id",
          "account_id",
          "region",
          "resource_type",
          "observed_at",
          "window_start",
          "window_end",
          "data_quality",
          "id"
        ],
        "title": "ResourceResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}